Add support for some i2c sensors and put them in rpicamsrc's annotation-text.
I've obtained a cheap bh1750fvi and sht3x and put them on the pi's i2c bus.

This will be used to automatically switch to night mode on darkness,
which will also enable some of the lighting automatically.

The lighting has to be controlled from here, because it affects the lux
measurement.